+++ change_log.md
@@ -5,6 +5,7 @@
 * New: Add mod button, automatically place mod in subfolder.
 * New: Dll mods may now receive a setting string, if a config file exists.
 * New: Dll mods may now export a setting class/struct and have the config file parsed by Modnix.
+* New: Dll mods may get modsRoot, modPath, and assemblyPath as initialiser params.
 * New: Modnix settings now stored in Modnix.conf which can be edited easily.
 * New: Formatted app/game/mod info.  Mod info panel now shows all display info, plus file list and entry points.
 * Fix: Mono.Cecil.dll is now deleted from Mods folder on setup, in addition to other loader dlls.